requirements.txt
python-frontmatter==1.1.0
PyYAML==6.0.1